chore: reviewcard 구조 개선
