+++ components/reviewCard/MyReviewCard.tsx
@@ -1,0 +1,3 @@
+export default function MyRivewCard() {
+  return;
+}
